<localization> binding
This PoC implements a XBL binding for <localization> elements adding a stub of
the Localization object API to them.  The <localization> elements are
referenced by localizable nodes and used by the mutation observer to get the
translation from the right resource.

In XUL, you can do:

    <localization id="browser-l10n" observe="main-window">
        <source src="chrome://branding/locale/brand.ftl"/>
        <source src="chrome://browser/locale/browser.ftl"/>
    </localization>

    <textbox id="urlbar"
             localization="browser-l10n"
             data-l10n-id"urlbar"
             …>

In XBL, you can do (the xul: namespace is wrong, but works for now):

    <xul:localization id="tabbrowser-l10n">
        <xul:source src="chrome://browser/locale/tabbrowser.ftl"/>
    </xul:localization>

    <![CDATA[
        var tab = document.createElementNS(NS_XUL, "tab");
        tab.setAttribute('localization', 'tabbrowser-l10n');
        tab.setAttribute('data-l10n-id', 'tabs-empty-tab-title');
    ]]>

The main difference between the two is the `observe` attribute that defines the
identifier of the root element to observe.  Unlike HTML's ShadowDOM, XBL's
anonymous content (AC) is visible from the outside, i.e. from non-anonymous
content (NAC).  Any mutation to the AC will be caught by the observer observing
the NAC root.  It also happens that XBL sets localizable attributes (e.g.
`label`) on nodes in NAC, but the translations have to come from the XBL's
resource.  A good example is the creation of new tabs:  <tabbrowser>'s `addTab`
method implmented in XBL (#tabbrowser) creates a XUL <tab> element, inserts it
into the main NAC XUL document, and binds the #tabbrowser-tab implementation.
The <tab>'s `label` can be a "New Tab" string which is defined in
tabbrowser.xml (XBL) resources, not in browser.xul.  By declaring the
`localization` attribute on each localizable node (in NAC and AC) we're making
sure that the observer knows which Localization object (and consequently, which
resources) should be used to translate this node.

Traversing the DOM tree top-down (e.g. when translateDocument(root) is called)
doesn't cross the boundary between NAC and AC.  Each <localization> element in
XBL will have to explicitly localize anonymous nodes on language change or
startup.  This isn't implemented in this PoC yet, but
document.getAnonymousNodes() could be helpful.

PoC disclaimer: The resources specified in <source> elements are not currently
fetched nor parsed at all.  I just return hardcoded translations for "New Tab"
and "Enter address or search" strings.