Resolver: Allow for no default value in Entity
Sometimes it's valid for an entity to not have a value and don't define
a default trait.  For instance, a DOM element might only require one of its
attributes to be localized:

    input =
      [placeholder] Placeholder

Localization.formatEntities already uses MessageContext.formatToPrimitive to
account for null values.  With this change formatToPrimitive also doesn't
report a "RangeError: No default" for null value which would trigger an
undesired fallback.